DeskTop: Consolidate padding.
DeskTop has a bunch of routines that are used by overlays. These need
to be placed in memory so they don't get overwritten, and were placed
after $A500. This led to padding to $A500 and later padding to $BF00.
There's nothing magical about $A500 though - there are just routines
that should be after $A000, and this can be handled with assertions
before those procs.

Drop the padding before $A500. This makes it easier to see that we
currently have... 2741 bytes of padding in main memory. Yay!