Show Text File: Decouple scroll position from line position
The scrollbar can only handle 0...255. To handle files with more
lines, have the first visible line be the canonical position in the
doc, and use the scrollbar as an input and visualization only. This
requires using mul/div functions with wider ranges.

Along the way, do a fair bit of cleanup and dead code removal too.

Fixes #700