Use correct branches after cmp16
The unsigned cmp16 macro uses the cmp/sbc approach, which only correctly
sets the C flag. The N flag depends on the high bit of the values
being compared, so it's valid for some range of inputs but not others
e.g. cmp16 #$8000, #$0001 would be wrong.

This correct usage in a handful of places which are obvious unsigned
comparisons. There are still some places in desktop main and aux that
use the N flag after the comparison, but these require careful
inspection to make sure they're truly unsigned comparisons and
shouldn't be made scmp16.

Tests for the Z flag after cmp16 are also sketchy and should be
looked at carefully.