DeskTop: Ensure desktop icons are cached during drag-select
The logic for dragging a selection rectangle on the desktop did not
ensure that desktop icons were cached. If a window's icons were cached
instead, this would result in a bogus selection and mis-painted icons.

This was made easier to repro by 52117ff8, but was possible to repro
before by clicking on a window's header area. The fix is to cache in
the desktop's icons before initiating the drag-select.