DeskTop: Fix "wandering" drag select rectangle within windows
When dragging a selection rectangle within a window, if the cursor
moved outside the window the rectangle would start to "wander"
within the window. This was caused by logic that tried to update
only the relevant edge of the rectangle each time.

Simplify by stashing the initial position, and building a rect
from that and the event coords.