DeskTop: Run DAs with MGTK's zero page preservation turned off.
Rather than flipping MGTK into zero page preservation mode before
running a DA, leave it alone. DAs that need the ZP preserved can turn
it on. This is done for Eyes, Calculator, and Sci.Calc, all of which
use the floating point routines in ROM and therefore depend on the
upper half of the ZP being preserved.

This means that DAs like Find Files and Control Panel need fewer hacks
to adjust the IP blink speed. Find Files needs logic to watch for
mouse moves to run at the same speed as other prompts. the Control
Panel is still doing more work in event processing so it gets a speed
adjustment, but less of one. Another benefit is that when DAs move
around the underlying desktop windows will repaint more quickly.