MGTK: Improve appearance of dimmed menu items
MGTK used a "light speckle" pattern as the mask for dimming menu
items; this is the same light speckle as used for the shaded scroll
bar regions. This results in some overly dark vertical lines when used
on text.

A plain checkerboard was used for disabled buttons, and this looks
better (based on a non-scientific poll), so apply to menus as well.