DeskTop: Improve performance of open-then-close-current.
When doing a "open the selected folder then close the current window"
action, the sequence is to open all the selected folders, then
activate the previous window, then close it. This is awkward
especially at 1MHz because due to overlaps the new windows usually end
up painting twice - once when they open, then again after the parent
is re-activated then closed. And the parent is repainted in the middle
of this, even though it's about to go away.

Shave off part of this time by not painting the contents of the
about-to-close parent window.

For #351