DeskTop: Further improve performance of open-then-close-current
See previous commit comment for more context.

Inspecting the MGTK CloseWindow call, it can close inactive windows.
Therefore, activating the parent window is not necessary. The child
windows will still paint twice (once on open, then as needed for the
update) but this is still be somewhat better.

Rework DeskTop's CloseWindow routine to handle closing non-active
windows as well, and take advantage of this in a couple of other
places that close arbitrary windows.

For #351