DeskTop: Move buttonloop and doubleclick libs from aux to main
The ButtonEventLoop and DetectDoubleClick procs are defined in libs.
Back when Main memory was tight these were instantiated in Aux, with
LC relays. Now we've freed up a ton of Main memory and Aux/LC are
tight, so move them to Main instead.

This is +291 bytes for Aux (bringing free space to 850 bytes), +18
bytes for LC, and and -291 bytes for Main, reducing headroom for
future code to "only" 2388 bytes. (Numbers from EN build.)