Selector: Don't trash DEVLST if launched w/o entries
In f501dd83 code was added to backup DEVLST, move the startup drive to
last, and then restore the original order on exit (e.g. when launching
DeskTop). Unfortunately, this backup step was skipped if DeskTop was
launched because the SELECTOR.LIST was empty, so DEVLST would be
trashed, causing DeskTop to hang.

Only restore the backup if it is valid.