File Dialog: Correct logic for updating Change Drive button status
Noticed via code inspection: a flag was incorrectly tested when
redrawing the Change Drive button if its status changed. This is moot
since the status can't change - it's based on the number of ProDOS
devices, which is static within the dialog's lifecycle - but might as
well make the code correct.