DeskTop: Fix initial window size for windows without icons.
Regressed in ca8a3da7 - that change was correct, but it revealed that
the icon bounding box calculation is bogus for windows without icons.
However, rather than tweaking that (because it's tricky) fix the code
that assumes it's valid to just use the min size if the window has no
icons.