DeskTop: Rearrange Format/Erase dialog to consolidate padding.
The Format/Erase dialog includes a Disk II format routine that must be
page aligned. Since this was in the middle of the overlay, this meant
that as the overlay was refactored padding to page-align it could
shift, making it a maintenance burden.

Simplify by moving the format routine to the start of the page-aligned
overlay, since the actual entry point from DeskTop into the overlay
doesn't need to be page aligned.