DeskTop: Prevent windows being dragged behind menu bar. Fixes #707
Changes in 5633583 removed logic that adjusted the window position
after a drag so that the window title bar did not appear behind the
menu bar, either intentionally or by accident - I don't recall.
However, if a window ends up with the title bar behind the menu, it
reveals an MGTK bug: when such a window is activated, the title bar
repaints on top of the menu bar.

Restore the fix.