MGTK: Fix painting of windows behind menu bar. Fixes #707 betterer.
This reverts the previous change and fixes the underlying bug in MGTK
instead: a non-clipped port is active when the window title is drawn
during activating a window, leading to the window frame painting over
the menu.

Re-assert the clipped port, and re-allow windows to be positioned
behind the menu bar. That behavior matches classic Mac OS, even though
you can't do it in contemporary macOS.