DeskTop: Move mul/div lib instantiation from aux to main
The 16-bit-mul/div routines are only used from main. Now that a lot of
space has been freed up in the main bank, relocate the lib
instantiation and drop the LC relay. Saves 18 bytes in LC, and shifts
~116 bytes from aux to main.

This leaves ~960 bytes free in aux. Done with an eye towards making
the line edit lib into a proper toolkit - it's currently ~640 bytes.