Introduce LineEdit ToolKit (LETK)
This reworks the line edit (text field) control logic from requiring
one instance of the code per control to a re-usable toolkit with an
MLI-style calling interface that operates on LineEditRecord instances.

The code gets somewhat more complicated, but this means that DAs don't
need their own copy of the library, and all instances inside DeskTop
-- rename (etc) dialog plus two in the file dialog -- can all share
the code.

The API itself could use some improvement (e.g. Activate should set
the blink_ip_flag) but this was intended to be a drop-in replacement
at first.

Related to #81.

No (intentional) behavior changes.