Line Edit: Fix bad IP location in Update
Update should be called when the string changes, but if the IP was
after the new string's length, badness could ensue (e.g. if there were
characters without glyphs after the end of the string). Clamp the IP
pos to be within the bounds of the string, just to be sure.