Introduce Button ToolKit (BTK)
This replaces lib/buttons.s with a new MLI-style interface, handling
tracking the mouse when a click is started and flashing the button
when the shortcut key is pressed. It adds a draw command to make
drawing a button a single call, and rolls in disable/enable as well
(currently used only in the file picker).

Like LETK, this means desk accessories don't need their own copy of
the library - they can call into DeskTop's copy of the TK. The LETK
files are moved from //letk/ into //toolkits/ to keep the top level
directory cleaner. MGTK is left alone (for now) since it's pretty
important.

No (intentional) behavior changes. In the future this could enable:

* Contextually disabled buttons in other dialogs.
* Drawing the shortcut label right aligned.
* A dialog manager to handle click and shortcut key dispatching.
* Radio and checkbox buttons could be supported as well.

Related to #81.