DeskTop: Shuffle resources to consolidate icon bitmaps
Move things around and especially move a 512-byte buffer from LC to
Main, leading to being able to place all of the file type icons
adjacent to each other in LC. Huzzah! And everything - Main, Aux, and
LC - are left with a fair bit of padding, even in l10n builds.