File Pickers: Eliminate pathname entry to simplify UX
Users are often confused how DeskTop's File > Copy a File... and File
> Delete a File... commands work, because they don't operate on the
selection. In addition, the file dialogs have a lot of clutter because
they allow the user to type a full pathname and behave strangely if
the selected path and typed path don't match.

Simplfy this by eliminating the pathname field from the dialogs.

* Selector > File > Run a Program...
  * Dialog reduced to just the picker UI.
* DeskTop > File > Copy a File...
  * Renamed to: "Copy To..."
  * Operates on selection.
  * Dialog reduced to just the picker UI.
* Desktop > File > Delete a File...
  * Renamed to: "Delete"
  * Operates on selection.
  * No dialog needed.
* Desktop > Shortcuts > Add/Edit a Shortcut...
  * Dialog only needs one text edit control.

Since the path is no longer shown, the picker dialogs now show the
current directory (volume or folder) name above the list box, and the
disk name is moved above the Change Drive button.

Other changes:

* When dragging files to the trash, the dialog no longer references
  emptying the trash.
* If the dialog doesn't have an edit field, type-down doesn't require
  holding a modifier key.

Resolves #699