DeskTop: Ensure Edit a Shortcut dialog behaves if target dir missing
When navigated to a directory / volume that's not present, the file
picker is supposed to just skip to the next volume, but the code for
this was convoluted and somewhat broken (possibly by refactors).
Simplify it. Also remove an `open_dir_flag` that doesn't seem to have
useful behavior, since the dialog will always have an open dir.