DeskTop: Fix updating internal paths when boot volume renamed
Bug in the logic for updating paths. If the startup path was "/HD/A2D"
and /HD was renamed to /HDX, the path was not correctly updated
because the "is this a substring?" test was checking if the character
after "/HD/A2D" was a slash, rather than testing if the character
after the "/HD" substring was a slash.

Already covered by the test plan, just need to actually run it.