Improve backslash character appearance in EN System font.
The backslash and forward slash were not symmetric - the backslash was
lower and thinner. Make it symmetric with the forward slash.

This only affects the EN System font; that code point is used for
other characters in non-EN languages, and the EN fixed width "Monaco"
font already had matching slashes.