Button ToolKit: Add flag for Draw to not set the port.
In Selector, if you ran File > Run a Program then dismissed the
dialog, the "update" process is used to refresh the window. As part of
that, the buttons are redrawn. But BTK was setting the full window
port, so the update clipping region was lost, causing repaints of more
of the dialog. And since those used an XOR brush, parts of the dialog
would disappear!

This only affected the `filedialogs` branch though, because in `main`
the file picker dialog completely covers the Selector dialog so it
gets a full repaint. But upstreaming the changes since they're general
goodness unrelated to the filedialogs change.

There's probably a better approach, but this works for now.