New desk accessory: CD Remote
Apple's release of the AppleCD SC in 1988 was accompanied by
supporting software for the Macintosh, the 16-bit Apple IIgs, and the
8-bit Apple IIe. This software included a "remote" for controlling
audio CD playback from the computer. In the case of the Apple IIe
software, it was a stand-alone ProDOS-8 program controlled using
keyboard commands.

@apple2geek disassembled the code over in
https://github.com/a2stuff/cdremote

This adapts the code into a DA, leaving most of the logic intact, just
swapping in a new interface.