Build: Add tool to generate an "LKG" package
LKG = "Last Known Good"

Run bin/lkg to do a build and generate packages. It will then copy the
2MG file, create a default config file, and tweak it to not copy to
RAMCard on startup. The result is in out/lkg.2mg. This is intended for
"demo" purposes, for example with Apple //jse[1]

This required adding bin/defopt to produce a default option file.

[1] https://www.scullinsteel.com/apple/e