DeskTop: Demote Screen Dump to Extras folder out of Apple menu
The Screen Dump is pretty esoteric - it dumps the screen to an
ImageWriter printer in Slot 1. It doesn't handle any other
configuration or errors.

Move it out of the Apple menu by default but leave it in Extras. This
gives users room to put one more DA of their choice in the Apple menu.
And making the DA a little more obscure will probably prevent
confusion in the future.