I18n: Use language decimal separator in Calculator and Sci.Calc DAs
In locales (well, languages) that use a decimal comma instead of a
decimal point, accept the comma key as a separator and show the comma
in calculation results. As a convenience, the '.' can still be typed.

L10n and i18n settings should really be decoupled, but that's a future
project.