Matrix: Use read operations for $C08x softswitches
The $C08x softswitches that control LC banking need to be accessed
with read operations. The Matrix DA was using write operations, which
kinda works but leaves the bank in a read-only mode. This is likely
the cause of #703, confirmation pending.

Switch them to using read operations (BIT works), and update
a few other places as well.