DeskTop: Fix prompt line edit control initialization
The proc used for initializing the line edit control in prompts was
missing an RTS at the end. Prior to 1d7a5858 this apparently was
harmless, but after that revision moved things around it caused
duplicate text to appear. Yay? Anyway, easy fix. It could have
been a lot worse!