Disk Copy: Don't adjust case for DOS 3.3 disk names
The logic to adjust case in volume names was executed in three
places:

* When drawing entries in the devices list.
* When showing the source name before a copy, if a ProDOS disk.
* When prompting about overriding the desination, if a ProDOS disk.

The first call site unconditionally adjusted the case, even if the
disk was not a ProDOS disk and had been synthesized e.g. "DOS 3.3
S6,D1", causing it to appear as "Dos 3.3".

To fix, move the case adjustment to the list population phase,
specifically in the code that deals only with ProDOS disks.