DeskTop: Show confirmation prompt when double-clicking BIN file.
In the original MouseDesk/A2D, you needed to hold down a modifier
to launch a BIN file. This was a safety measure to make sure you
really meant it, as a non-executable would crash the computer.

In 1.2 we softened this a bit and allowed running using the menu
commands (File > Open or Apple-O) without the safety measure, but
double-click was still confusing for users.

Since this is such a common issue, add an alert to confirm the
action, rather than just silently failing.