DeskTop: Introduce a file (vs folder) icon in file list views
In DeskTop's directory list views, as well as the file list for file
picker dialogs, include an icon for files, not just folders. Only one
icon type is used - a document.

Also, tweak the spacing metrics for the name and icons in list views
to make them consistent.

This uses up two glyphs in the fonts, but the glyphs being replaced
weren't useful or used anyway.