DeskTop: Fix corruption/crash after closing a window
Logic added to cause the active window's icon to be selected after a
click (if nothing else was) didn't handle the case when the window was
closed by the click, leaving no active window. It caused subtle memory
corruption that was only noticed by @apple2geek when testing later
changes in Apple //jse - we haven't found another easy repro.