DeskTop: Show keyboard shortcut for File > Delete in the menu
Rather than adding custom handling and making users guess at the
shortcut, just show it in the menu.

The only gotcha is... what glyph to use? Unfortunately, the canonical
backspace glyph ⌫ looks really bad in a 7-pixel wide font. And since
that wasn't used "back in the day" anyway... just use the symbol you
get when you type Delete at the Applesoft prompt: ▒

Also, tweak the font glyph to move it up to the baseline.