IconTK: Skip unnecessary work when (un)highlighting volume icons during drag
When dragging icons over drop targets, the targets are
highlighted/unhighlighted on the fly. Adjusted ports are calculated
for icons so that the highlighting is clipped. This isn't necessary
for desktop icons, which have their own port setting/clipping logic.
So skip adjustments in these cases, which should be a minor perf
boost.