IconTK: Make DrawAll draw in order, not reverse order
Introduced in e17563b2 the new IconTK::DrawAll call takes care of
rendering all icons in a window (or desktop). But it iterated the
icons in reverse order, which at 1MHz makes windows fill backwards.
This isn't wrong, but it's disconcerting.

Flip the iteration order to match DeskTop's previous behavior.