DeskTop: Remove some redundant scrollbar init code
Remove some old code responsible for initializing scrollbars in
directory windows. The max values can be defined at build time.
Actually enabling the scrollbar and setting the position is done by
the `RedrawAfterContentChange` call after `MGTK::OpenWindow` is called
in `OpenWindowForIcon`. All that's needed is to set the scrollbars to
a present-but-inactive state in case the winfo struct was previously
used.