DeskTop: Save/restore view type across restarts
As mentioned in 988b88f8, now that list views are not useless it makes
sense to save and restore the view type in addition to window bounds
across restarts.

The LOCAL/DESKTOP.FILE file format version is incremented, so you'll
lose window state when updating to this version (or downgrading, if
you're playing git shennanigans).

Fit and finish following #28