This Apple DA: Fix memory display edge case on Apple IIgs
When computing memory, slots are enumerated looking for slinky memory
card firmware signatures. The loop checked slots 7..0 (not 7..1) -
which meant softswitches on the $C0xx page were being read. Ooof.

This manifested (under Ample/MAME) on the IIgs as showing "000,000"
for memory, probably because something critical was banked out. Very
lucky this didn't cause crashes. Thanks to testers for finding this.