DeskTop: Persist window scroll position across restarts
Previously, what was stored was the window position and dimensions.
Now the window position (viewloc) and viewport (maprect) are
persisted. This simplifies the code, too.

So this doesn't actually store the scroll position, just the
*viewport*, or the view into the window's coordinate space. The scroll
position is derived from this but since icon positions are not
persisted (please don't ask for that) if you move the icons the actual
scrollbar positions might not be restored.

The LOCAL/DESKTOP.FILE file format version is incremented, so you'll
lose window state when updating to this version (or downgrading, if
you're playing git shennanigans).

Fit and finish following #28