IconTK: Don't clamp drag highlight to screen bounds
This behavior dating back to the original code felt really bad. If you
were dragging icons, the outlines were clamped to the screen. This
meant if you got near the edge of the screen the highlight would
"disconnect" from the mouse which is awkward. If you're dragging a lot
of icons it happens almost immediately.

So... just delete the complex code that implements it. Saves a
whopping 663 bytes. Yay!