IconTK: Enable clipped rendering for windowed icons
* IconTK::DrawIcon is renamed to IconTk::DrawIconRaw indicating it
  relies on the caller to set up a grafport. Additionally, it no
  longer clips volume icons on the desktop.

* IconTK::DrawIcon is new, and always clips, and handles clipping both
  volume icons and windowed icons against overlapping windows. The
  trickiest part is ensuring that the shading pattern is offset
  correctly.

From DeskTop's perspective, this offers a simpler API - if a single
icon needs to be updated, just use IconTK::DrawIcon. No need to fuss
around with GrafPorts. It also means that operations on icons that
might be in background windows are now possible. For now, the only
change is that dimmed icons (for open folders) can change state
appropriately now. In the future, supporting selection in background
windows may be possible.

This saves about 60 bytes in Main, and costs 380 bytes in Aux. But we
earned it.