DeskTop: Enable dropping files on folders in inactive windows
Now that icons in inactive windows can be painted, why not allow
dropping in them? This turned out to be relatively easy! The only
gotcha is that MGTK's FindControl - which is used to ensure that the
mouse is over a client portion of the window, not the scrollbars -
only works on the topmost window. So, add FindControlEx (awesome name,
right?) that can work on any window.

There are probably some edge cases that'll turn up, but it seems to
work so far!