Date & Time DA: Use BTK for the OK button
Date & Time DA uses bespoke logic for handling clicks on the date/time
fields and up/down arrows. It also used this for the OK button. This
meant that niceties of the button such as the rendering of the label
with shortcut flush right and tracking during a click weren't handled.

Now that BTK is "free" for DAs, use it, just for the OK button.