DeskTop: Collapse window entry and icon usage tables, free aux space
DeskTop reserved 128-byte buffers for the desktop and each of 8
possible windows in Aux memory ($1B00 through $1F7F) to hold the list
of icons they contained. This was just over 1k of memory. Since only
127 icons are supported, each of these buffers was mostly empty space.
Since Aux memory is difficult to use, all access to these buffers was
via cache load/save procs in precious LC memory. This change collapses
those buffers into a single 128-byte buffer in Main memory, and moves
the load/save functions into Main as well. A table of offsets for each
window is used, and when saving back the list the entries are shifted
and offsets are updated as appropriate.

DeskTop also used Aux $1F80-$1FFF for a used/free icon map, with
another pair of LC routines for allocating/freeing icons. This change
moves that map and the procs into main.

This frees $1B00-$1FFF (1.25k) in Aux. It's not super useful space,
but it does simplify Desk Accessories which were free to use through
$1C00 in Main during load and through $2000 after, but only through
$1B00 in Aux. Otherwise, the space is declared as more "save area" for
MGTK usage.

* Main: 469 bytes consumed (more than half are the tables)
* LC: 157 bytes freed in $D000-$D200 (the relay proc block)
* Aux: 1280 bytes freed under $2000