DeskTop: Don't move icons when dragged over source window's non-content area.
Dropping an icon on another window's title bar, scroll bars/grow box,
or header should copy (or move) the file. But on the same window it
should be a no-op, but wasn't.

Refactor the code used to clip hit testing of icons to the real
content area to short-circuit same-window moves.