MGTK: Only expand pattern when it has changed
SetPatternImpl takes the 1-byte/8-row pattern and "expands" it to an
80-byte by 8-row pattern, ready for easy masking onto the screen. This
is slow - about 16000 cycles - and SetPatternImpl is called whenever
the GrafPort changes. When moving selection from one icon to another,
this accounts for ~10% of the cycles.

Optimize by tracking the last used pattern, and only perform the
expansion if the pattern has actually changed. Since many draw
operations end up being into the same GrafPort, and many ports have
the same default pattern anyway, this avoids unnecessary work.