DeskTop: Fix formatting disks with controllers in all slots.
The logic for testing the $CnFE byte to see if a driver supports
formatting was flawed, so the driver entry point plus $FE was checked
instead (e.g. $C764 + $FE), giving bogus results. This would lead A2D
to believe that to controllers (like Virtual ]['s Omnidisks) that
don't support the FORMAT command actually did, try it, and fail the
format.

Fix the logic.