DeskTop: Improve performance of window drawing and selection clearing
Icons are stored in screen coordinates, to make dragging easier. When
redrawing a window (and many other operations), a window's icons are
mapped from screen to window, one by one. This was done by looping
over all icons in the window and calling a mapping proc. The mapping
proc in turn called a proc to compute the window/screen offsets. That
meant this proc was called redundantly for every single icon in the
loop.

Factor this offset proc out of the "for all icons in the window" and
"for all selected icons" loops.