DeskTop: Fix initial window height when not min/max out
Regressed in 66decf80. While the horizontal sizing code was changed to
be smart and use the boundng box, the vertical sizing code was not.
But the bbox initialization was changed for both.

Revert the vertical bbox initialization. This asymmetry makes the code
a little more difficult to understand, but it's the simplest fix.