IIgs: Slow processor when playing alert sounds
Since the IIgs doesn't auto-slow when audio is playing, slow the
processor manually around the playback routine. The tricky bit here is
that the ROM/LCBANK state is not known by the routine, so the caller
of the Bell proc is required to have pre-probed and expose an
is_iigs_flag.