Sounds DA: Add padding to ensure proc swap doesn't break routines
When playing a sound, the sound proc is byte-swapped into $300 (the
playback location). And for convenience, we don't track the length of
each routine - the maximum length is used. This is fine except when a
routine used by the swap (e.g. JTRelay) is part of the swap range, and
therefore gets swapped out. Oops.

Add padding - it's overkill, but safe.

This has always been risky but became a problem in 73147884.