NORMFAST: Scope screen hole poking to Laser 128 code path
Back in cf89dacb the NORMFAST code was modified to poke a screen hole
at $7FE which is needed on the Laser 128 to make the speed change
sticky. This was done unconditionally, and it does bad things on the
IIgs.

Change the code to only do the screen hole mod on the Laser 128.