DeskTop: Reset window viewport after repopulating.
Following 66decf80 the initial icon layout for a window doesn't
include left padding, and instead relies on the window viewport being
adjusted based on the icon bounding box. This didn't account for cases
where the window's icons are refreshed.

Factor out the logic and call it in the SelectAndRefresh path too.