Build: Remove a bunch of unreferenced symbols/code
Temporarily assemble with -W9 and look for any "real" unused symbols.
Unfortunately, param blocks (either inline or via macros) generate a
ton of unused symbols so this isn't something we can leave on.

That said, it pointed out that Selector was still including the
LineEdit TK on behalf of the file dialog, which is no longer needed
for that usage, which shaves 1k off the module. So it's worth
doing this as a sanity check every so often.