Include ProDOS clock driver in packages
This is the new "jumbo" driver from over at:

https://github.com/a2stuff/prodos-drivers/tree/main/clocks/jumbo

This includes support for the No-Slot Clock, ROMX, FujiNet, AE DClock,
and The Cricket!; it's silent on success or failure, and not
necessarily what users will want to install to their hard disk, but it
should make running off of the pre-made images a little better.

The file is a snapshot, so we'll need to update as the "jumbo" driver
(or those it pulls in) evolves.