DeskTop: Tweak dialog label spacing
Previously, the item spacing used in DeskTop's common dialogs was the
same as the font height. (Well, prior to 1.2 it was one pixel shorter
than the font height, but we fixed that.) This meant that descenders
would touch the top of capital letters.

Make the spacing one more, to improve readability. This affects the
following dialogs:

* Apple > About Apple II DeskTop
* File > New Folder
* File > Get Info
* File > Rename...
* File > Duplicate...
* File > Copy To...
* File > Delete...
* Special > Format a Disk...
* Special > Erase a Disk...
* Special > Lock...
* Special > Unlock...
* Special > Get Size

... and when files are being copied to RAMCard.