Build: Name the module DESKTOP instead of DESKTOP2
DESKTOP2 is a legacy from MouseDesk when the launcher SYS file was
named DESKTOP1, as a hint to users to run that first. Subsequent
hobbyists renamed it to DESKTOP.SYSTEM to make it even more obvious
that it should be launched. Especially now that the files are stuffed
into the MODULES/ subdirectory, drop the "2".