Another attempt to clear text screen when exiting DHR mode.
The change in 0a2b9ea3 is mostly correct, unless the text window has
not been set. Calling HOME without resetting the window won't
necessarily clear the screen. Calling INIT will cause the text screen
to briefly show.

Drop the call to INIT entirely and manually set the text window to
0,0,80,24 before calling HOME.

For #621 and #639