DeskTop: Change JUMP_TABLE_LAUNCH_FILE to be useful
There's an unused (for now) JT API that launches a file. But... the
file it operates on requires the path to the directory and the name to
exist in DeskTop-internal buffers, so it's useless from DAs.

Switch the JT entry to use a slightly different entry point that takes
the path in a well-known buffer instead.