Invoker: Move early SET_PREFIX call after interpreter branch
When launching a file directly (no interpreter), the prefix needs to
be set as the Invoker is passed a prefix path and a relative path. But
the interpreter has a full path and needs to set the prefix after the
interpreter is loaded.

Avoid a redundant call (but not save any bytes) by reordering the code
slightly.