DeskTop: Avoid polling empty drives twice on startup
DeskTop polls all drives (via an ON_LINE call) to populate the volume
icons. When restoring windows, it tries to open windows for each saved
path. If the path is for a volume that isn't on-line (e.g. an ejected
disk) this will cause a re-poll by ProDOS as it looks for the path.
This is especially irritating for noisy Disk II drives.

Since DeskTop knows what volumes are on-line, check each path before
trying to restore the window and skip if needed.