Text Preview DA: Handle long files more intelligently
Previously, the Show.Text.File accessory's algorithm for rendering was
to set a viewport based on scroll position, then start at the top of
the file and render downwards until the bottom of the viewport was
hit. This meant that as you scrolled down a file, time was spent
rendering content "above" the viewport for each re-display. For a very
long file this was a terrible experience even under unbounded
emulation, and completely impractical at 1MHz.

To address this, special-case the first display of the file, and file
offsets for lines in the file. For a long file, this means that the
initial load is followed by a delay (with a watch cursor) while the
rest of the file is parsed. But afterwards, scrolling to any position
is snappy. To handle long files, not every line's offset is recorded
(as that would chew up memory). Instead, every Nth (N=16) line
position is recorded; this means some time is spent during redisplay
on lines "above" the viewport, but far less than before.

As a side effect, this also renders the heuristics in 180ef368 (for
the file in lines is known, so the scrollbar positioning is now
precise.

Some dead code and unnecessary buffers are also removed.