Desktop: Format/Erase - Default device to selected volume
One of the complaints about DeskTop[1] is that some commands are
difficult to use because of the reliance on slot/drive specifications.
One place this manifests is the Special > Format a Disk... and Special
> Erase a Disk... commands that prompt the user for a device, listing
all the units with the type of device (e.g. Disk II, etc) and
Slot/Drive, which may not be obvious especially given ProDOS's
remapping.

It's hard to move away from this entirely, because DeskTop can't
unobtrusively scan all drives. The early Mac could force you to choose
between formatting a disk (therefore giving it a name) or ejecting it,
but given the lack of file system translators we don't want to do that
in DeskTop.

As a partial mitigation: if a single volume icon is selected, and
Format/Erase a Disk is invoked, skip the device picker and operate on
the specified volume's device.

[1] https://guidebookgallery.org/articles/apple2userinterfaces