Neko: Move behavior and most UI logic into main to save space
Neko was using 98% of the aux space - holding both resources and
behavior logic - so there was no room for any additions. By moving the
behavior logic into main, there's now plenty of overhead in both.

After this change, only 26% of main and 78% of aux is in use.

With this extra room, erasing when moving is now optimized to just the
previous frame's bounds.