DeskTop: Draw header before entries when refreshing window
The header used to rely on the icon count to give the "items" count,
which meant that when repopulating a window (e.g. after a delete,
duplicate, drop, etc) the items would draw before the header - which
looks kinda weird at 1MHz.

Use the window's file record count instead, and draw the header first.
Less weird!