Desktop System: Fix self-update for path code
When DeskTop starts up, DESKTOP.SYSTEM stashes the "original" prefix
in LC memory, right after the ProDOS Selector (a.k.a. QUIT handler) it
installs. This is used so that if running from a RAMCard and something
needs updating (shortcuts, settings, etc) the system can update not
just the copy on the RAMCard but the original system disk.

It also has logic to write this path back to a header block in
DESKTOP.SYSTEM itself so that if DeskTop is restarted from RAMCard it
still has that path. Neat, huh?

Unfortunately... the code to write it was broken, with an incorrect
path. Easy fix.